Add playGame
Moved score variables and playRound function into
playGame function. Made a for loop for 5 rounds
that calls the choices for the human and computer
each round so it doesn't use the same values for
the duration of the game. In the end the function
displays the total scores of each player and determines
overall winner. This way we can know our own scores
to make sure the end call was correct.